review of enigme activity is completed
