+++ src/pages/creer-une-enigme/2.tsx
@@ -2,7 +2,7 @@ import { useRouter } from 'next/router';
 import React from 'react';
 
 import { isEnigme } from 'src/activity-types/anyActivity';
-import { ENIGME_DATA, ENIGME_TYPES } from 'src/activity-types/enigme.constants';
+import { ENIGME_DATA, ENIGME_TYPES, getEnigme } from 'src/activity-types/enigme.constants';
 import type { EnigmeData } from 'src/activity-types/enigme.types';
 import { Base } from 'src/components/Base';
 import { StepsButton } from 'src/components/StepsButtons';
@@ -18,7 +18,7 @@ const EnigmeStep2 = () => {
   const { activity, updateActivity, addContent, deleteContent, save } = React.useContext(ActivityContext);
 
   const data = (activity?.data as EnigmeData) || null;
-  const indiceContentIndex = Math.max(data?.indiceContentIndex ?? 0, 0);
+  const indiceContentIndex = data?.indiceContentIndex ?? 0;
 
   const errorSteps = React.useMemo(() => {
     if (data !== null) {
@@ -27,32 +27,53 @@ const EnigmeStep2 = () => {
     return [];
   }, [data]);
 
+  const contentAdded = React.useRef(false);
   React.useEffect(() => {
     if (activity === null && !('activity-id' in router.query) && !sessionStorage.getItem('activity')) {
       router.push('/creer-une-enigme');
     } else if (activity && !isEnigme(activity)) {
       router.push('/creer-une-enigme');
     }
-  }, [activity, router]);
+
+    if (activity && isEnigme(activity)) {
+      if ((activity.data.indiceContentIndex ?? 0) > activity.content.length) {
+        updateActivity({
+          data: {
+            ...activity.data,
+            indiceContentIndex: activity.content.length,
+          },
+        });
+      }
+      if ((activity.data.indiceContentIndex ?? 0) === activity.content.length && !contentAdded.current) {
+        contentAdded.current = true;
+        addContent('text');
+      }
+    }
+  }, [activity, router, updateActivity, addContent]);
 
   const updateContent = (content: ActivityContent[]): void => {
     if (!activity) {
       return;
     }
-    updateActivity({ content: [...content, ...activity.content.slice(indiceContentIndex, activity.content.length)] });
+    updateActivity({ content: [...activity.content.slice(0, indiceContentIndex), ...content] });
   };
-  const addDescriptionContent = (type: ActivityContentType, value?: string) => {
-    addContent(type, value, indiceContentIndex);
-    updateActivity({ data: { ...data, indiceContentIndex: indiceContentIndex + 1 } });
+  const addIndiceContent = (type: ActivityContentType, value?: string) => {
+    contentAdded.current = true;
+    addContent(type, value);
   };
-  const deleteDescriptionContent = (index: number) => {
-    deleteContent(index);
-    updateActivity({ data: { ...data, indiceContentIndex: indiceContentIndex - 1 } });
+  const deleteIndiceContent = (index: number) => {
+    contentAdded.current = true; // delete means there were content already
+    deleteContent(indiceContentIndex + index);
   };
 
   if (data === null || activity === null || !isEnigme(activity)) {
     return <div></div>;
   }
+
+  const onNext = () => {
+    save().catch(console.error);
+    router.push('/creer-une-enigme/3');
+  };
 
   const enigmeType = ENIGME_TYPES[activity.subType ?? 0] ?? ENIGME_TYPES[0];
   const enigmeData = ENIGME_DATA[activity.subType ?? 0] ?? ENIGME_DATA[0];
@@ -62,9 +83,13 @@ const EnigmeStep2 = () => {
       <div style={{ width: '100%', padding: '0.5rem 1rem 1rem 1rem' }}>
         <Steps
           steps={[
-            data.theme === -1 ? capitalize(data.themeName ?? '') : enigmeData[data.theme]?.step ?? 'Choix de la catégorie',
-            enigmeType.step1 ?? "Description de l'objet",
-            "Création de l'indice",
+            data.theme === -1
+              ? capitalize(data.themeName ?? '')
+              : activity.subType === -1
+              ? getEnigme(activity.subType, data).step1
+              : enigmeData[data.theme]?.step ?? 'Thème',
+            'Énigme',
+            'Réponse',
             'Prévisualisation',
           ]}
           urls={['/creer-une-enigme/1?edit', '/creer-une-enigme/2', '/creer-une-enigme/3', '/creer-une-enigme/4']}
@@ -72,20 +97,19 @@ const EnigmeStep2 = () => {
           errorSteps={errorSteps}
         />
         <div className="width-900">
-          <h1>{enigmeType.titleStep2}</h1>
+          <h1>Créez l’énigme, pour faire deviner votre objet</h1>
           <p className="text" style={{ fontSize: '1.1rem' }}>
-            Décrivez ici votre {enigmeType.titleStep2Short}, il s’agira de la <strong>réponse</strong> partagée aux autres classes. Votre réponse ne
-            sera visible que 7 jours après la publication de votre énigme, pour laisser le temps à vos Pélicopains de faire des recherches, et de vous
-            poser des questions !
+            Faites deviner votre {enigmeType.titleStep2Short} à vos Pélicopains en donnant des indices en vidéos, images, sons et texte !<br></br> À
+            l’étape suivante, vous écrirez la réponse à l’énigme.
           </p>
           <ContentEditor
-            content={activity.content.slice(0, indiceContentIndex)}
+            content={activity.content.slice(indiceContentIndex, activity.content.length)}
             updateContent={updateContent}
-            addContent={addDescriptionContent}
-            deleteContent={deleteDescriptionContent}
+            addContent={addIndiceContent}
+            deleteContent={deleteIndiceContent}
             save={save}
           />
-          <StepsButton prev={`/creer-une-enigme/1?edit=${activity.id}`} next="/creer-une-enigme/3" />
+          <StepsButton prev={`/creer-une-enigme/1?edit=${activity.id}`} next={onNext} />
         </div>
       </div>
     </Base>
