+++ src/pages/creer-une-enigme/3.tsx
@@ -2,7 +2,7 @@ import { useRouter } from 'next/router';
 import React from 'react';
 
 import { isEnigme } from 'src/activity-types/anyActivity';
-import { ENIGME_DATA, ENIGME_TYPES } from 'src/activity-types/enigme.constants';
+import { ENIGME_DATA, getEnigme } from 'src/activity-types/enigme.constants';
 import type { EnigmeData } from 'src/activity-types/enigme.types';
 import { Base } from 'src/components/Base';
 import { StepsButton } from 'src/components/StepsButtons';
@@ -18,68 +18,47 @@ const EnigmeStep3 = () => {
   const { activity, updateActivity, addContent, deleteContent, save } = React.useContext(ActivityContext);
 
   const data = (activity?.data as EnigmeData) || null;
-  const indiceContentIndex = data?.indiceContentIndex ?? 0;
+  const indiceContentIndex = Math.max(data?.indiceContentIndex ?? 0, 0);
 
   const errorSteps = React.useMemo(() => {
-    const fieldStep2 = activity?.content.filter((d) => d.value !== ''); // if value is empty in step 2
+    const fieldStep2 = activity?.content.filter((d) => d.value !== '' && d.value !== '<p></p>\n'); // if value is empty in step 2
     if (data !== null && fieldStep2?.length === 0) {
-      const errors = getErrorSteps(data, 2);
+      const errors = getErrorSteps(data, 1);
       errors.push(1); //corresponding to step 2
       return errors;
     }
-    if (data !== null) return getErrorSteps(data, 2);
+    if (data !== null) return getErrorSteps(data, 1);
+
     return [];
   }, [activity?.content, data]);
 
-  const contentAdded = React.useRef(false);
   React.useEffect(() => {
     if (activity === null && !('activity-id' in router.query) && !sessionStorage.getItem('activity')) {
       router.push('/creer-une-enigme');
     } else if (activity && !isEnigme(activity)) {
       router.push('/creer-une-enigme');
     }
-
-    if (activity && isEnigme(activity)) {
-      if ((activity.data.indiceContentIndex ?? 0) > activity.content.length) {
-        updateActivity({
-          data: {
-            ...activity.data,
-            indiceContentIndex: activity.content.length,
-          },
-        });
-      }
-      if ((activity.data.indiceContentIndex ?? 0) === activity.content.length && !contentAdded.current) {
-        contentAdded.current = true;
-        addContent('text');
-      }
-    }
-  }, [activity, router, updateActivity, addContent]);
+  }, [activity, router]);
 
   const updateContent = (content: ActivityContent[]): void => {
     if (!activity) {
       return;
     }
-    updateActivity({ content: [...activity.content.slice(0, indiceContentIndex), ...content] });
+    updateActivity({ content: [...content, ...activity.content.slice(indiceContentIndex, activity.content.length)] });
   };
-  const addIndiceContent = (type: ActivityContentType, value?: string) => {
-    contentAdded.current = true;
-    addContent(type, value);
+  const addDescriptionContent = (type: ActivityContentType, value?: string) => {
+    addContent(type, value, indiceContentIndex);
+    updateActivity({ data: { ...data, indiceContentIndex: indiceContentIndex + 1 } });
   };
-  const deleteIndiceContent = (index: number) => {
-    contentAdded.current = true; // delete means there were content already
-    deleteContent(indiceContentIndex + index);
+  const deleteDescriptionContent = (index: number) => {
+    deleteContent(index);
+    updateActivity({ data: { ...data, indiceContentIndex: indiceContentIndex - 1 } });
   };
 
   if (data === null || activity === null || !isEnigme(activity)) {
     return <div></div>;
   }
 
-  const onNext = () => {
-    save().catch(console.error);
-    router.push('/creer-une-enigme/4');
-  };
-
-  const enigmeType = ENIGME_TYPES[activity.subType ?? 0] ?? ENIGME_TYPES[0];
   const enigmeData = ENIGME_DATA[activity.subType ?? 0] ?? ENIGME_DATA[0];
 
   return (
@@ -87,9 +66,13 @@ const EnigmeStep3 = () => {
       <div style={{ width: '100%', padding: '0.5rem 1rem 1rem 1rem' }}>
         <Steps
           steps={[
-            data.theme === -1 ? capitalize(data.themeName ?? '') : enigmeData[data.theme]?.step ?? 'Choix de la catégorie',
-            enigmeType.step1 ?? "Description de l'objet",
-            "Création de l'indice",
+            data.theme === -1
+              ? capitalize(data.themeName ?? '')
+              : activity.subType === -1
+              ? getEnigme(activity.subType, data).step1
+              : enigmeData[data.theme]?.step ?? 'Thème',
+            'Énigme',
+            'Réponse',
             'Prévisualisation',
           ]}
           urls={['/creer-une-enigme/1?edit', '/creer-une-enigme/2', '/creer-une-enigme/3', '/creer-une-enigme/4']}
@@ -97,19 +80,20 @@ const EnigmeStep3 = () => {
           errorSteps={errorSteps}
         />
         <div className="width-900">
-          <h1>Créer votre indice</h1>
+          <h1>Créez la réponse à votre énigme</h1>
           <p className="text" style={{ fontSize: '1.1rem' }}>
-            Créez ici un <strong>indice</strong> pour faire deviner votre {enigmeType.titleStep2Short}. Vous pouvez ajouter du texte, une vidéo ou une
-            image à votre indice et vous pourrez le modifier à l’étape 4.
+            À présent, écrivez la réponse à votre énigme, en donnant des explications à vos Pélicopains. Votre réponse ne sera visible que{' '}
+            <strong>7 jours</strong> après la publication de votre énigme, pour laisser le temps à vos Pélicopains de faire des recherches, et de vous
+            poser des questions !
           </p>
           <ContentEditor
-            content={activity.content.slice(indiceContentIndex, activity.content.length)}
+            content={activity.content.slice(0, indiceContentIndex)}
             updateContent={updateContent}
-            addContent={addIndiceContent}
-            deleteContent={deleteIndiceContent}
+            addContent={addDescriptionContent}
+            deleteContent={deleteDescriptionContent}
             save={save}
           />
-          <StepsButton prev="/creer-une-enigme/2" next={onNext} />
+          <StepsButton prev={`/creer-une-enigme/2`} next="/creer-une-enigme/4" />
         </div>
       </div>
     </Base>
